+++ packages/contracts/src/framework/plugin/repo/PluginRepo.sol
@@ -60,9 +60,6 @@ contract PluginRepo is
     /// @notice The mapping between the plugin setup address and its corresponding version hash.
     mapping(address => bytes32) internal latestTagHashForPluginSetup;
 
-    /// @notice The mapping between release id and release's metadata URI.
-    mapping(uint8 => bytes) internal metadataPerRelease;
-
     /// @notice The ID of the latest release.
     /// @dev The maximum release number is 255.
     uint8 public latestRelease;
@@ -71,9 +68,8 @@ contract PluginRepo is
     /// @param versionHash The tag hash.
     error VersionHashDoesNotExist(bytes32 versionHash);
 
-    /// @notice Thrown if a contract does not inherit from `PluginSetup`.
-    /// @param invalidPluginSetup The address of the contract missing the `PluginSetup` interface.
-    error InvalidPluginSetupInterface(address invalidPluginSetup);
+    /// @notice Thrown if a plugin setup contract does not inherit from `PluginSetup`.
+    error InvalidPluginSetupInterface();
 
     /// @notice Thrown if a release number is zero.
     error ReleaseZeroNotAllowed();
@@ -89,14 +85,11 @@ contract PluginRepo is
     /// @param pluginSetup The plugin setup contract address.
     error PluginSetupAlreadyInPreviousRelease(uint8 release, uint16 build, address pluginSetup);
 
-    /// @notice Thrown if the metadata URI is not set for a release.
-    /// @param release the release number for which the metadata URI is not set.
-    /// @param releaseMetadata The release metadata URI.
-    error InvalidReleaseMetadata(uint8 release, bytes releaseMetadata);
+    /// @notice Thrown if the metadata URI is empty.
+    error EmptyReleaseMetadata();
 
     /// @notice Thrown if release does not exist.
-    /// @param release The release number of the release that does not exist.
-    error ReleaseDoesNotExist(uint8 release);
+    error ReleaseDoesNotExist();
 
     /// @notice Thrown if the same plugin setup exists in previous releases.
     /// @param release The release number.
@@ -138,7 +131,7 @@ contract PluginRepo is
         bytes calldata _releaseMetadata
     ) external auth(address(this), MAINTAINER_PERMISSION_ID) {
         if (!_pluginSetup.supportsInterface(type(IPluginSetup).interfaceId)) {
-            revert InvalidPluginSetupInterface({invalidPluginSetup: _pluginSetup});
+            revert InvalidPluginSetupInterface();
         }
 
         if (_release == 0) {
@@ -154,10 +147,7 @@ contract PluginRepo is
             latestRelease = _release;
 
             if (_releaseMetadata.length == 0) {
-                revert InvalidReleaseMetadata({
-                    release: _release,
-                    releaseMetadata: _releaseMetadata
-                });
+                revert EmptyReleaseMetadata();
             }
         }
 
@@ -188,7 +178,7 @@ contract PluginRepo is
         });
 
         if (_releaseMetadata.length > 0) {
-            _updateReleaseMetadata(_release, _releaseMetadata);
+            emit ReleaseMetadataUpdated(_release, _releaseMetadata);
         }
     }
 
@@ -202,23 +192,14 @@ contract PluginRepo is
         }
 
         if (_release > latestRelease) {
-            revert ReleaseDoesNotExist({release: _release});
+            revert ReleaseDoesNotExist();
         }
 
         if (_releaseMetadata.length == 0) {
-            revert InvalidReleaseMetadata({release: _release, releaseMetadata: _releaseMetadata});
+            revert EmptyReleaseMetadata();
         }
 
-        _updateReleaseMetadata(_release, _releaseMetadata);
-    }
-
-    /// @notice The private helper function to replace new `_metadata` for the `_release`.
-    /// @param _release The release number.
-    /// @param _metadata The release metadata URI.
-    function _updateReleaseMetadata(uint8 _release, bytes calldata _metadata) internal {
-        metadataPerRelease[_release] = _metadata;
-
-        emit ReleaseMetadataUpdated(_release, _metadata);
+        emit ReleaseMetadataUpdated(_release, _releaseMetadata);
     }
 
     /// @notice Returns the latest version for a given release number.